tests now use float64 by default; some reformatting was performed

+++ src/problem.jl
@@ -44,9 +44,9 @@ The equation to be solved is of the form:
 # Examples
 ```julia
 # Free propagation example
-L = 10.0
+L = 10
 N = 128
-u0 = (rand(ComplexF32, N, N),)
+u0 = (rand(ComplexF64, N, N),)
 dispersion(ks, param) = sum(abs2, ks) / 2
 prob = GrossPitaevskiiProblem(u0, (L, L); dispersion)
 ```
@@ -56,20 +56,20 @@ prob = GrossPitaevskiiProblem(u0, (L, L); dispersion)
 
 # Define system parameters
 param = (;
-    ħ = 1.0,           # Reduced Planck constant
-    m = 1.0,           # Effective mass
-    δc = 0.0,          # Cavity detuning
-    δx = 0.0,          # Exciton detuning
+	ħ = 1,           # Reduced Planck constant
+    m = 1,           # Effective mass
+    δc = 0,          # Cavity detuning
+    δx = 0,          # Exciton detuning
     γc = 0.1,          # Cavity decay rate
     γx = 0.1,          # Exciton decay rate
-    Ωr = 2.0,          # Rabi coupling
+    Ωr = 2,          # Rabi coupling
     g = 0.01,          # Nonlinearity strength
-    A = 1.0,           # Pump amplitude
-    w = 2.0            # Pump width
+    A = 1,           # Pump amplitude
+    w = 2            # Pump width
 )
 
 # Define grid parameters
-L = 10.0
+L = 10
 N = 128
 lengths = (L, L)
 
@@ -87,7 +87,7 @@ end
 nonlinearity(u, param) = @SVector [0, param.g * abs2(u[2])]
 
 function pump(r, param, t)
-    SVector(param.A * exp(-sum(abs2, r) / param.w^2), 0.0)
+    SVector(param.A * exp(-sum(abs2, r) / param.w^2), 0)
 end
 
 # Create the problem
@@ -95,45 +95,45 @@ prob = GrossPitaevskiiProblem(u0, lengths; dispersion, nonlinearity, pump, param
 ```
 """
 struct GrossPitaevskiiProblem{N,M,T1,T2,T3,T4,T5,T6,T7,T8,T9,T10}
-  u0::NTuple{M,T1}
-  lengths::NTuple{N,T2}
-  dispersion::T3
-  potential::T4
-  nonlinearity::T5
-  pump::T6
-  position_noise_func::T7
-  momentum_noise_func::T8
-  noise_prototype::T9
-  param::T10
-  function GrossPitaevskiiProblem(u0::Tuple, lengths::Tuple; dispersion::T3=additiveIdentity, potential::T4=additiveIdentity,
-    nonlinearity::T5=additiveIdentity, pump::T6=additiveIdentity,
-    position_noise_func::T7=additiveIdentity, momentum_noise_func::T8=additiveIdentity, noise_prototype::T9=additiveIdentity,
-    param::T10=nothing) where {T3,T4,T5,T6,T7,T8,T9,T10}
-    @assert all(x -> ndims(x) ≥ length(lengths), u0)
-    @assert all(x -> size(x) == size(first(u0)), u0)
-    _u0 = complex.(u0)
-    _lengths = promote(lengths...)
-    N = length(_lengths)
-    M = length(_u0)
-    T1 = eltype(_u0)
-    T2 = typeof(first(_lengths))
-    new{N,M,T1,T2,T3,T4,T5,T6,T7,T8,T9,T10}(_u0, _lengths, dispersion, potential, nonlinearity,
-      pump, position_noise_func, momentum_noise_func, noise_prototype, param)
-  end
+    u0::NTuple{M,T1}
+    lengths::NTuple{N,T2}
+    dispersion::T3
+    potential::T4
+    nonlinearity::T5
+    pump::T6
+    position_noise_func::T7
+    momentum_noise_func::T8
+    noise_prototype::T9
+    param::T10
+    function GrossPitaevskiiProblem(u0::Tuple, lengths::Tuple; dispersion::T3=additiveIdentity, potential::T4=additiveIdentity,
+        nonlinearity::T5=additiveIdentity, pump::T6=additiveIdentity,
+        position_noise_func::T7=additiveIdentity, momentum_noise_func::T8=additiveIdentity, noise_prototype::T9=additiveIdentity,
+        param::T10=nothing) where {T3,T4,T5,T6,T7,T8,T9,T10}
+        @assert all(x -> ndims(x) ≥ length(lengths), u0)
+        @assert all(x -> size(x) == size(first(u0)), u0)
+        _u0 = complex.(u0)
+        _lengths = promote(lengths...)
+        N = length(_lengths)
+        M = length(_u0)
+        T1 = eltype(_u0)
+        T2 = typeof(first(_lengths))
+        new{N,M,T1,T2,T3,T4,T5,T6,T7,T8,T9,T10}(_u0, _lengths, dispersion, potential, nonlinearity,
+            pump, position_noise_func, momentum_noise_func, noise_prototype, param)
+    end
 end
 
 function Base.show(io::IO, ::GrossPitaevskiiProblem{N}) where {N}
-  print(io, "$(N)D GrossPitaevskiiProblem")
+    print(io, "$(N)D GrossPitaevskiiProblem")
 end
 
 direct_grid(L, N) = StepRangeLen(zero(L), L / N, N)
 
 function direct_grid(prob::GrossPitaevskiiProblem)
-  map(direct_grid, prob.lengths, size(first(prob.u0)))
+    map(direct_grid, prob.lengths, size(first(prob.u0)))
 end
 
 reciprocal_grid(L, N) = fftfreq(N, oftype(N / L, 2π) * N / L)
 
 function reciprocal_grid(prob::GrossPitaevskiiProblem{N}) where {N}
-  map(reciprocal_grid, prob.lengths, size(first(prob.u0)))
+    map(reciprocal_grid, prob.lengths, size(first(prob.u0)))
 end
